requirements.txt
################################################################################
# This requirements files has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

-e .
apistar-shell==1.0
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
django==2.0.2
idna==2.6
ipython-genutils==0.2.0
ipython==6.2.1
itypes==1.1.0
jedi==0.11.1
jinja2==2.10
markupsafe==1.0
parso==0.1.1
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pygments==2.2.0
pytest==3.4.0
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
werkzeug==0.14.1
whitenoise==3.3.1
